Source Code Of Humilator-V3/requirements.txt
aiohttp
discord.py
discord.py-self
requests
colorama
discord_webhook
numpy
pywin32
pycryptodome
